Address review concerns
KAFKA-257
