+++ src/main/java/com/mongodb/kafka/connect/sink/dlq/WriteException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mongodb.kafka.connect.sink;
+package com.mongodb.kafka.connect.sink.dlq;
 
 import java.util.Formatter;
 import java.util.Locale;
@@ -26,10 +26,10 @@ import com.mongodb.WriteError;
  * org.bson.BsonDocument#toJson()}. We may change it in the future, in which case the version
  * (marked with {@code v}) will be incremented.
  */
-final class WriteException extends NoStackTraceDlqException {
+public final class WriteException extends NoStackTraceDlqException {
   private static final long serialVersionUID = 1L;
 
-  WriteException(final WriteError error) {
+  public WriteException(final WriteError error) {
     super(
         String.format(
             Locale.ENGLISH,
