+++ src/main/java/com/mongodb/kafka/connect/sink/dlq/package-info.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2008-present MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Contains classes related to {@linkplain
+ * org.apache.kafka.connect.sink.ErrantRecordReporter#report(org.apache.kafka.connect.sink.SinkRecord,
+ * java.lang.Throwable) reporting to the dead letter queue (DLQ)}.
+ */
+package com.mongodb.kafka.connect.sink.dlq;
