+++ src/main/java/com/mongodb/kafka/connect/util/Assertions.java
@@ -18,9 +18,9 @@ package com.mongodb.kafka.connect.util;
 import com.mongodb.lang.Nullable;
 
 /**
- * All {@code assert...} and {@code fail} methods throw {@link AssertionError} and should be used to
- * check conditions which may be violated if and only if the driver code is incorrect. The intended
- * usage of these methods is the same as of the <a
+ * All methods throw {@link AssertionError} and should be used to check conditions which may be
+ * violated if and only if the driver code is incorrect. The intended usage of these methods is the
+ * same as of the <a
  * href="https://docs.oracle.com/javase/8/docs/technotes/guides/language/assert.html">Java {@code
  * assert} statement</a>. The reason for not using the {@code assert} statements is that they are
  * not always enabled. We prefer having internal checks always done at the cost of our code doing a
@@ -28,20 +28,6 @@ import com.mongodb.lang.Nullable;
  * values to open possibilities of being used fluently.
  */
 public final class Assertions {
-  /**
-   * @param value A value to check.
-   * @param <T> The type of {@code value}.
-   * @return {@code null}.
-   * @throws AssertionError If {@code value} is not {@code null}.
-   */
-  @Nullable
-  public static <T> T assertNull(@Nullable final T value) throws AssertionError {
-    if (value != null) {
-      throw new AssertionError(value.toString());
-    }
-    return null;
-  }
-
   /**
    * @param value A value to check.
    * @param <T> The type of {@code value}.
@@ -65,37 +51,6 @@ public final class Assertions {
       throw new AssertionError();
     }
     return true;
-  }
-
-  /**
-   * @param value A value to check.
-   * @return {@code false}.
-   * @throws AssertionError If {@code value} is {@code true}.
-   */
-  public static boolean assertFalse(final boolean value) throws AssertionError {
-    if (value) {
-      throw new AssertionError();
-    }
-    return false;
-  }
-
-  /**
-   * @throws AssertionError Always
-   * @return Never completes normally. The return type is {@link AssertionError} to allow writing
-   *     {@code throw fail()}. This may be helpful in non-{@code void} methods.
-   */
-  public static AssertionError fail() throws AssertionError {
-    throw new AssertionError();
-  }
-
-  /**
-   * @param msg The failure message.
-   * @throws AssertionError Always
-   * @return Never completes normally. The return type is {@link AssertionError} to allow writing
-   *     {@code throw fail("failure message")}. This may be helpful in non-{@code void} methods.
-   */
-  public static AssertionError fail(final String msg) throws AssertionError {
-    throw new AssertionError(assertNotNull(msg));
   }
 
   private Assertions() {}
